requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.in
#
aiosqlite==0.17.0
    # via databases
alembic==1.8.1
    # via -r requirements.in
anyio==3.6.1
    # via starlette
astroid==2.11.7
    # via pylint
bcrypt==3.2.2
    # via passlib
black==22.6.0
    # via -r requirements.in
cffi==1.15.1
    # via
    #   bcrypt
    #   cryptography
click==8.1.3
    # via
    #   black
    #   uvicorn
cryptography==37.0.4
    # via python-jose
cuid==0.3
    # via -r requirements.in
databases[aiosqlite]==0.6.0
    # via -r requirements.in
dill==0.3.5.1
    # via pylint
ecdsa==0.18.0
    # via python-jose
fastapi==0.79.0
    # via -r requirements.in
greenlet==1.1.2
    # via sqlalchemy
h11==0.12.0
    # via uvicorn
idna==3.3
    # via anyio
invoke==1.7.1
    # via -r requirements.in
isort==5.10.1
    # via
    #   -r requirements.in
    #   pylint
lazy-object-proxy==1.7.1
    # via astroid
mako==1.2.1
    # via alembic
markupsafe==2.1.1
    # via mako
mccabe==0.7.0
    # via pylint
mypy==0.961
    # via -r requirements.in
mypy-extensions==0.4.3
    # via
    #   black
    #   mypy
passlib[bcrypt]==1.7.4
    # via -r requirements.in
pathspec==0.9.0
    # via black
platformdirs==2.5.2
    # via
    #   black
    #   pylint
pyasn1==0.4.8
    # via
    #   python-jose
    #   rsa
pycparser==2.21
    # via cffi
pydantic==1.9.1
    # via
    #   fastapi
    #   sqlmodel
pylint==2.14.5
    # via -r requirements.in
python-dotenv==0.20.0
    # via -r requirements.in
python-jose[cryptography]==3.3.0
    # via -r requirements.in
python-multipart==0.0.5
    # via -r requirements.in
rsa==4.8
    # via python-jose
six==1.16.0
    # via
    #   ecdsa
    #   python-multipart
sniffio==1.2.0
    # via anyio
sqlalchemy==1.4.39
    # via
    #   alembic
    #   databases
    #   sqlmodel
sqlalchemy2-stubs==0.0.2a24
    # via sqlmodel
sqlmodel==0.0.6
    # via -r requirements.in
starlette==0.19.1
    # via fastapi
tomli==2.0.1
    # via
    #   black
    #   mypy
    #   pylint
tomlkit==0.11.1
    # via pylint
typing-extensions==4.3.0
    # via
    #   aiosqlite
    #   mypy
    #   pydantic
    #   sqlalchemy2-stubs
uvicorn==0.18.2
    # via -r requirements.in
wrapt==1.14.1
    # via astroid

# The following packages are considered to be unsafe in a requirements file:
# setuptools
